requirements.txt
gradio>=4.31
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0.76
gfpgan>=1.3.8
basicsr>=1.4.2
facexlib>=0.3.0
